+++ packages/react/src/index.ts
@@ -67,5 +67,6 @@ export * from "./display";
 export * from "./map";
 export * from "./preview";
 export * from "./previewFixtures";
+export * from "./streetGeometryPreview";
 export * from "./saveSlots";
 export * from "./cameraShake";